feat: module artilces

+++ src/common/base/base.service.ts
@@ -12,8 +12,8 @@ export class BaseService<T extends BaseEntity, R extends Repository<T>>
     this.repository = repository;
   }
 
-  findAll(): Promise<T[]> {
-    return this.repository.find();
+  async findAll(): Promise<T[]> {
+    return await this.repository.find();
   }
 
   async findById(id: EntityId): Promise<T> {
@@ -22,20 +22,22 @@ export class BaseService<T extends BaseEntity, R extends Repository<T>>
     return result;
   }
 
-  findByIds(ids: [EntityId]): Promise<T[]> {
-    return this.repository.findByIds(ids);
+  async findByIds(ids: [EntityId]): Promise<T[]> {
+    return await this.repository.findByIds(ids);
   }
 
-  create(data: any): Promise<T> {
-    return this.repository.save(data);
+  async create(data: any): Promise<T> {
+    return await this.repository.save(data);
   }
 
   async update(id: EntityId, data: any): Promise<T> {
     await this.repository.update(id, data);
-    return this.findById(id);
+    return await this.findById(id);
   }
 
-  remove(id: EntityId): Promise<DeleteResult> {
-    return this.repository.delete(id);
+  async remove(id: EntityId): Promise<DeleteResult> {
+    const result = await this.repository.findOne(id);
+    if (!result) throw new NotFoundException();
+    return await this.repository.delete(id);
   }
 }
